feat: keep user-space tun as an option
